Pin the parent Devfile version to 1.2.0
The default version of the parent Stack will be updated in [1] to use JDK 17,
so that users on recent versions of Spring Boot can use such Stack.

This is to make sure this sample does not break with such change,
and continues to work as before, since it is using Sprint Boot 2.x (and
expects JDK 11).

Signed-off-by: Armel Soro <asoro@redhat.com>